+++ CHANGELOG.md
@@ -8,6 +8,15 @@
 
 - The time widget is now editable when replay is paused
 
+## 0.4.1
+
+### Bug Fixes
+
+- The time widget is now editable when replay is paused
+- After importing a CSV file only pause when in connected state
+- Do not start the elapsed timer when skipping to begin of recording
+- Properly hide window when stay on top is deselected
+
 ## 0.4.0
 
 ### New Features